lab 4/app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
mysql-connector-python==9.0.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.66
requests>=2.31.0
statsmodels==0.14.2
scikit-learn==1.5.1
pyarrow==16.1.0
